Rename files in API error package
Now the names more closely correspond to the actual behavior
