+++ src/main/java/com/md/scheduler/configuration/api/errors/ApiErrorDetails.java
@@ -1,7 +1,8 @@
-package com.md.scheduler.configuration.baseApiExceptions;
+package com.md.scheduler.configuration.api.errors;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 
+
 @JsonInclude(JsonInclude.Include.NON_NULL)
-interface ApiSubException {
+interface ApiErrorDetails {
 }
