+++ src/main/java/com/md/scheduler/configuration/api/errors/EntityNotFoundDetails.java
@@ -1,11 +1,11 @@
-package com.md.scheduler.configuration.baseApiExceptions;
+package com.md.scheduler.configuration.api.errors;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-class ApiEntityNotFoundException implements ApiSubException {
+class EntityNotFoundDetails implements ApiErrorDetails {
 
     private final String field;
     private final Object notFoundValue;
