+++ src/main/java/com/md/scheduler/configuration/api/errors/ValidationErrorDetails.java
@@ -1,21 +1,21 @@
-package com.md.scheduler.configuration.baseApiExceptions;
+package com.md.scheduler.configuration.api.errors;
 
 import lombok.Getter;
 
 @Getter
-class ApiValidationException implements ApiSubException {
+class ValidationErrorDetails implements ApiErrorDetails {
 
     private String object;
     private String field;
     private Object rejectedValue;
     private String message;
 
-    public ApiValidationException(String object, String message) {
+    public ValidationErrorDetails(String object, String message) {
         this.object = object;
         this.message = message;
     }
 
-    public ApiValidationException(String object, String field, Object rejectedValue, String message) {
+    public ValidationErrorDetails(String object, String field, Object rejectedValue, String message) {
         this.object = object;
         this.field = field;
         this.rejectedValue = rejectedValue;
